Move logging middleware to middleware package
