+++ core/src/main/scala/zio/web/http/HttpMiddleware.scala
@@ -2,12 +2,6 @@ package zio.web.http
 
 import java.nio.charset.StandardCharsets
 import zio._
-import zio.blocking.Blocking
-import zio.clock._
-import zio.stream.{ Sink, ZSink, ZStream }
-
-import java.io.{ FileOutputStream, IOException, OutputStream }
-import java.time.format.DateTimeFormatter
 
 /**
  * An `HttpMiddleware[R, E]` value defines HTTP middleware that requires an
@@ -92,48 +86,6 @@ object HttpMiddleware {
         )
       )
     )
-
-  def logging[R, E](sink: ZSink[R, E, String, Byte, Long]): HttpMiddleware[Clock with R, Nothing] = {
-    val formatter = DateTimeFormatter.ofPattern("dd/MMM/yyyy:HH:mm:ss Z")
-
-    HttpMiddleware(
-      for {
-        queue  <- Queue.bounded[String](128)
-        stream = ZStream.fromQueue(queue)
-        _      <- stream.run(sink).fork
-      } yield Middleware(
-        request(HttpRequest.Method.zip(HttpRequest.URI).zip(HttpRequest.Version).zip(HttpRequest.IpAddress)) {
-          case (((method, uri), version), ipAddr) =>
-            val ipStr = ipAddr.fold("-")(_.getHostAddress)
-            currentDateTime
-              .fold(
-                _ => s"$ipStr - - - ${"\""}$method $uri $version${"\""}",
-                now => {
-                  val time = now.format(formatter)
-                  s"$ipStr - - [$time] ${"\""}$method $uri $version${"\""}"
-                }
-              )
-        },
-        Response(
-          HttpResponse.StatusCode.zip(HttpResponse.Header("Content-Length")),
-          (state: String, resp: (Int, String)) => queue.offer(s"$state ${resp._1} ${resp._2}\n").as(Patch.empty)
-        )
-      )
-    )
-  }
-
-  def fileSink(path: String): ZSink[Blocking, Throwable, String, Byte, Long] = {
-    val stream: ZManaged[Blocking, IOException, OutputStream] =
-      ZManaged
-        .fromAutoCloseable(
-          ZIO.effect(new FileOutputStream(path, true))
-        )
-        .refineToOrDie[IOException]
-
-    Sink
-      .fromOutputStreamManaged(stream)
-      .contramapChunks[String](_.flatMap(str => Chunk.fromIterable(str.getBytes)))
-  }
 
   def rateLimiter(n: Int): HttpMiddleware[Any, None.type] =
     HttpMiddleware(
